Update README to reflect Apache branding
I think the first heading with "Solr" in it should clearly show as "Apache Solr".  After the heading and first sentence, the remaining reference needn't.  I'm not certain if the page was a problem as it was (in the eyes of the ASF) but it should be a-okay now.  In addition, I added a reference to the ASF governance... this is after all a key differentiator between Solr and it's competitors.