Switch back to fetching keys from a keyserver, but use multiple keyservers to handle edge cases better
Also, this modifies `update.sh` to ensure that several keyservers definitely have a copy of the public key we need.

This should account for flakiness in ha.pool.sks-keyservers.net, as well as flakiness in environments where the default HKP port is blocked (by explicitly trying to use other keyservers that listen on port 80).