Make internal solr port configurable
We do not need to explicitly configure the SOLR_PORT in the solr.in.sh if we supply the SOLR_PORT as default Docker environment variable. Furthermore, we make the internal SOLR_PORT configurable.

This is useful during integration tests in cloud mode when used inside a docker-compose network. In those cases, the Solr connects to the Zookeeper using it’s internal node name (overwritable) and internal port number (now overwriteable).