Add centos based Solr images
In addition to `alpine`, `slim` and `debian` based distros this adds a `centos` based Docker-Solr image. This adds a new template, adjusts tests to handle centos distros as well.

As there is no ready made openjdk-centos image, a recent jdk is installed in the Dockerfile. Moreover the Java version can be specified via environment variable. Also the centos base image can be changed as Docker build argument. This can be handy, if you need a specific centos based image that uses the `yum` package manager.