Create a /opt/mysolrhome
This is a convenience for users using the INIT_SOLR_HOME=yes option
along with a docker volume.
Normally, if you create a volume and mount it at a non-existing
location /mysolrhome, it ends up owned by root, and you cannot
write to it as solr. Because we create the /opt/mysolrhome at build
time and make it owned by solr:solr, when you mount a volume there
you can write to that volume. For example:

    docker volume create --name mysolrhome
    docker run -v "mysolrhome:/opt/mysolrhome" -e SOLR_HOME=/opt/mysolrhome -e INIT_SOLR_HOME=yes solr:7.3 solr-demo