multistage build
Use multistage builds in the Dockerfiles, so we can avoid
build-specific things like gpg ending up in the image.

Refactor some bash code inlines in the Dockerfiles into
scripts in a separate builder_tools dir, so we can just
copy that to the builder images and run it. This makes
it easier to maintain, we can run shellceck, and share
the scripts between Dockerfiles.

Move the gosu fetching/checking into a separate builder.
This improves caching during development

No longer try to pull the base image prior to the docker build,
because the multiple FROMs have made that too complicated.

Stop printing the solr version, to further simplify the Dockerfile
that previously rewrote that. If we want that, we can do the
rewriting in update.sh

Remove deprecated MAINTAINER tag from the Dockerfiles,
use labels instead.

Move some ENVs to ARGs to reduce environmental pollution.