Add tini
To run with tini under Docker, just run with --init to
get the built-in one.
If your engine does not support something similar,
set TINI=yes to run Solr under the tini in the image.

I've moved the actual running of Solr from solr-foreground
to solr-fg, and invoke that from the other scripts, so they
get the consistent behaviour.