Update Readme for solrdata permissions (#364)
I added a command the to show that the solrdata directory on the host must have the proper permissions to be accessible by the solr user inside the container. In this case, I have done so by changing the user on the host to the same UID as the solr user in the container.