Update Dockerfile templates to delete unneeded dependencies
This updates the work @janhoy did in
0b13c14c7ae9190044b5d14c1268ba7f49a4dc33 by adjusting to a change in
shell behaviour: now that dash is the default shell, the brace expansion
does not work as expected but the `-f` flag passed to `rm` meant that
the build continued without deleting the intended files.

This partially addresses
https://issues.apache.org/jira/browse/SOLR-15855 because the removed
files have a number of open CVEs, several of which are serious and may
block deployments if used in a CI/CD pipeline which has a scanner.

Because dash does not support brace expansion and find’s `-delete`
action does not support deleting directories, this commit splits the
work into two commands: `rm -R` for the directories and `find … -delete`
for the files.